+++ scripts/importGitCoinProjects.ts
@@ -1,6 +1,6 @@
 import { AlchemyProvider } from "@ethersproject/providers";
 import { Space } from "@prisma/client";
-import { getProjectRegistryContract } from "lib/gitcoin/contracts";
+import { getProjectRegistryContract } from "lib/gitcoin/getProjectRegistryContract";
 import { getProjectDetails, GitcoinProjectDetails } from "lib/gitcoin/getProjectDetails";
 import { prisma } from 'db';
 import { uid } from 'lib/utilities/strings';
@@ -17,21 +17,6 @@ import { getFilenameWithExtension } from "lib/utilities/getFilenameWithExtension
  * NOTE: This script creates new users and spaces for Gitcoin projects.
  * It also updates mixpanel profiles so make sure to have prod mixpanel api key set in .env
  */
-
-export type GitcoinChainId = 1 | 5 | 10 | 250;
-
-export const GITCOIN_SUPPORTED_CHAINS = [1, 5, 10, 250] as const;
-
-export const PROJECT_REGISTRY_ADDRESSES: Record<GitcoinChainId, string> = {
-  // MAINNET
-  1: '0x03506eD3f57892C85DB20C36846e9c808aFe9ef4',
-  // OPTIMISM
-  10: '0x8e1bD5Da87C14dd8e08F7ecc2aBf9D1d558ea174',
-  // FANTOM
-  250: '0x8e1bD5Da87C14dd8e08F7ecc2aBf9D1d558ea174',
-  // GOERLI
-  5: '0x832c5391dc7931312CbdBc1046669c9c3A4A28d5'
-};
 
 const START_ID = 715;
 const CHAIN_ID = 1;
